Add handling for adding and  removing listings from the watchlist
It allow user to add and remove listing from thier watchlist.In the "add_to_watchlist" view, implemented fuctionality to check if the listing is already present in the user's watchlist. If the listing exists, it is removed from the watchlist. If not, the listing is added to the watchlist.